requirements.txt
fastapi~=0.111.0
SQLAlchemy~=2.0.30
pydantic~=2.7.1
python-dotenv~=1.0.1
uvicorn~=0.29.0
alembic~=1.13.1
fastapi-filter~=2.0.0
